Create model package inside core.items package and update imports
