+++ src/main/kotlin/core/mission/LootService.kt
@@ -3,6 +3,8 @@ package dev.deadzone.core.mission
 import dev.deadzone.core.data.assets.ItemResource
 import dev.deadzone.module.Dependency
 import dev.deadzone.core.data.GameResourceRegistry
+import dev.deadzone.core.mission.model.LootContent
+import dev.deadzone.core.mission.model.LootParameter
 import org.w3c.dom.Document
 import org.w3c.dom.Element
 import java.io.StringReader
@@ -26,25 +28,6 @@ val ALL_LOCS = listOf(
     "body", "basement", "ammo", "wood", "metal", "cloth", "comms", "hospital",
     "island", "firstaid"
 ) // + the tutorial convenience store (tutorialfuel)
-
-data class LootContent(
-    val itemId: String,
-    val itemIdInXML: String,
-    val quantity: Int,
-)
-
-// Loot system config, contains variables that affects loot
-data class LootParameter(
-    val areaLevel: Int,
-    val playerLevel: Int, // not used yet
-    val itemWeightOverrides: Map<String, Double> = emptyMap(), // absolute override if needed
-    val specificItemBoost: Map<String, Double> = emptyMap(),   // e.g. "fuel" to 0.2 (+20%)
-    val itemTypeBoost: Map<String, Double> = emptyMap(),       // e.g. "weapon" to 0.1 (+10%)
-    val itemQualityBoost: Map<String, Double> = emptyMap(),    // e.g. "blue" to -0.4 (-40%)
-    val baseWeight: Double = 1.0,
-    val fuelLimit: Int = 0 // not used yet
-)
-
 
 class LootService(
     private val gameResourceRegistry: GameResourceRegistry = Dependency.gameResourceRegistry,
