+++ src/main/kotlin/core/mission/model/LootContent.kt
@@ -1,0 +1,7 @@
+package dev.deadzone.core.mission.model
+
+data class LootContent(
+    val itemId: String,
+    val itemIdInXML: String,
+    val quantity: Int,
+)
